列表条目 展示

+++ app/src/main/java/com/lz/framecase/bean/PositionEntity.kt
@@ -4,14 +4,14 @@ package com.lz.framecase.bean
  * -----------作者----------日期----------变更内容-----
  * -          刘泽      2018-08-29       创建class
  */
- class PositionEntity {
+class PositionEntity {
     /**
      * latitude : 0
      * longitude : 0
      * position :
      */
 
-    var latitude: Int = 0
-    var longitude: Int = 0
+    var latitude: Double = 0.0
+    var longitude: Double = 0.0
     var position: String? = null
 }
